refactor(brett): extract auth.ts (OIDC, requireAdmin, boardAuthRedirect)
